기본과제완료
