--- packages/assignment/src/@lib/hocs/deepMemo.ts
@@ -1,9 +1,0 @@
-import { deepEquals } from "../equalities";
-import { ComponentType } from "react";
-import { memo } from "./memo.ts";
-
-export function deepMemo<P extends object>(
-  Component: ComponentType<P>,
-) {
-  return memo(Component, deepEquals);
-}
